Major refactor with docs update - Refactored code to elimited package loop repeats - Refactored for readability - Moved some config to `package.json` - Added instructions to package README.md file - Removed need for `update-docs` script, everything now runs within the docs consolidator
